+++ app/services/statistics_export_service.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'abstract_controller'
 require 'action_controller'
 require 'action_view'
@@ -13,7 +15,7 @@ class StatisticsExportService
     # query all stats with range arguments
     query = MultiJson.load(export.query)
 
-    @results = Elasticsearch::Model.client.search({index: 'stats', scroll: '30s', body: query})
+    @results = Elasticsearch::Model.client.search(index: 'stats', scroll: '30s', body: query)
     scroll_id = @results['_scroll_id']
     while @results['hits']['hits'].size != @results['hits']['total']
       scroll_res = Elasticsearch::Model.client.scroll(scroll: '30s', scroll_id: scroll_id)
@@ -22,9 +24,9 @@ class StatisticsExportService
     end
 
     ids = @results['hits']['hits'].map { |u| u['_source']['userId'] }
-    @users = User.includes(:profile).where(:id => ids)
+    @users = User.includes(:profile).where(id: ids)
 
-    @indices = StatisticIndex.all.includes(:statistic_fields, :statistic_types => [:statistic_sub_types])
+    @indices = StatisticIndex.all.includes(:statistic_fields, statistic_types: [:statistic_sub_types])
 
     ActionController::Base.prepend_view_path './app/views/'
     # place data in view_assigns
@@ -37,10 +39,10 @@ class StatisticsExportService
 
     content = av.render template: 'exports/statistics_global.xlsx.axlsx'
     # write content to file
-    File.open(export.file,"w+b") {|f| f.puts content }
+    File.open(export.file, 'w+b') { |f| f.puts content }
   end
 
-  %w(account event machine project subscription training space).each do |path|
+  %w[account event machine project subscription training space].each do |path|
     class_eval %{
       def export_#{path}(export)
 
@@ -76,7 +78,7 @@ class StatisticsExportService
         # write content to file
         File.open(export.file,"w+b") {|f| f.puts content }
       end
-    }
+    }, __FILE__, __LINE__ - 35
   end
 
 end