+++ db/migrate/20190730085826_add_extension_to_export.rb
@@ -1,0 +1,5 @@
+class AddExtensionToExport < ActiveRecord::Migration
+  def change
+    add_column :exports, :extension, :string, default: 'xlsx'
+  end
+end